requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
openai==1.3.7
python-dotenv==1.0.0
